+++ NEWS.md
@@ -8,6 +8,7 @@
 - `geom_miss_point()` works with `shape` argument #290
 - fix bug with `all_complete`, which was implemented as `!anyNA(x)` but should be `all(complete.cases(x))`.
 - correctly implement `any_na()` (and `any_miss()`) and `any_complete()`. Rework examples to demonstrate workflow for finding complete variables.
+- implement `Date`, `POSIXct` and `POSIXlt` methods for `impute_below()` - #158
 
 ## Misc